feat: sign-up functionality implemented
This commit adds a new hook called useAuthContext to the AuthContext file. The hook allows components to access the authentication context using the useContext hook from React. It throws an error if used outside of the AuthProvider component. This hook will be used to manage user authentication throughout the application.